Nuew contrsct uploaded

+++ contracts/Market.sol
@@ -14,21 +14,21 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 /// @dev All function calls are currently implemented without side effects
 contract Market {
     //Length of a frame
-    uint64 public period; //64b -- Data packing into storage slots (1 slot = 32 bytes) --> 4*64b = 32 bytes
+    uint public period; //64b -- Data packing into storage slots (1 slot = 32 bytes) --> 4*64b = 32 bytes
     //Contract creation time
-    uint64 public initTimestamp; //64b
+    uint public initTimestamp; //64b
     //Protocol fee
-    uint64 public feeProtocol = 100; //64b
+    uint public feeProtocol = 100; //64b
     //Reporting time for pair price
-    uint64 public tReporting; //64b
+    uint public tReporting; //64b
     //Token pair on Uniswap
     IUniswapV2Pair public uniswapPair; //256b
     //Range of a lot
-    uint256 public dPrice; //256b
+    uint public dPrice; //256b
     //Fee that gets to market owner
-    uint256 public feeMarket; //256b
+    uint public feeMarket; //256b
     //Tax on lots
-    uint256 public taxMarket; //256b
+    uint public taxMarket; //256b
     //Contract for creating markets
     MarketFactory public factory; //256b
     //ERC20 token used for buyng lots in this contract 
@@ -36,26 +36,26 @@ contract Market {
     //Owner of this market
     address public ownerMarket; //256b
     //Minimal lot tax per frame
-    uint256 public minTax; //256b
+    uint public minTax; //256b
     //Scalar number used for calculating precentages
-    uint256 scalar = 1e24; //256b
+    uint scalar = 1e24; //256b
 
     enum SFrame {NULL, OPENED, CLOSED, INVALID}
     struct Frame {
-        uint64 frameKey; 
-        uint64 oracleTimestampStart;
-        uint64 oracleTimestampEnd;
-        uint64 lastBlockNum; 
-        uint256 priceAverage;
-        uint256 rewardFund;
-        uint256 oraclePrice0CumulativeStart;
-        uint256 oraclePrice0CumulativeEnd;
+        uint frameKey; 
+        uint oracleTimestampStart;
+        uint oracleTimestampEnd;
+        uint lastBlockNum; 
+        uint priceAverage;
+        uint rewardFund;
+        uint oraclePrice0CumulativeStart;
+        uint oraclePrice0CumulativeEnd;
         SFrame state;
         uint[] lotKeys;
         mapping(uint => Lot) lots;
     }
 
-    event FrameUpdate(uint64 frameKey);
+    event FrameUpdate(uint frameKey);
 
     mapping(uint => Frame) public frames;
     mapping(address => uint[]) public userFrames;
@@ -63,15 +63,15 @@ contract Market {
 
     enum SLot {NULL, BOUGHT, SETTLED}
     struct Lot {
-        uint256 lotKey;
+        uint lotKey;
         address lotOwner;
-        uint256 acquisitionPrice;
+        uint acquisitionPrice;
         SLot state;
     }
 
-    event LotUpdate(uint64 frameKey, uint lotKey);
+    event LotUpdate(uint frameKey, uint lotKey);
 
-    constructor(MarketFactory _factory, IERC20 _token, IUniswapV2Pair _uniswapPair, uint64 _period, uint64 _initTimestamp, uint _taxMarket, uint _feeMarket, uint _dPrice, uint64 _tReporting, uint _minTax) {
+    constructor(MarketFactory _factory, IERC20 _token, IUniswapV2Pair _uniswapPair, uint _period, uint _initTimestamp, uint _taxMarket, uint _feeMarket, uint _dPrice, uint _tReporting, uint _minTax) {
         accountingToken = _token;
         ownerMarket = payable(msg.sender);
         period = _period;
@@ -111,7 +111,7 @@ contract Market {
     /// @notice Calculate frames timestamp (beggining of frame)
     /// @param timestamp In seconds
     /// @return frame's timestamp (key)
-    function clcFrameTimestamp(uint64 timestamp) public view returns (uint64){
+    function clcFrameTimestamp(uint timestamp) public view returns (uint){
         if (timestamp <= initTimestamp) return initTimestamp;
         return timestamp - ((timestamp - (initTimestamp)) % (period));
     }
@@ -142,22 +142,22 @@ contract Market {
     /// @notice Calculate how many frames there are left untill given frame in input
     /// @param frameKey Timestamp of an arbitary frame in the future
     /// @return number of frames (whole numbers: n=0,1,2...)
-    function clcFramesLeft(uint64 frameKey) public view returns (uint64){                      
-        return (((frameKey - (uint64(block.timestamp))) / period) + 1);
+    function clcFramesLeft(uint frameKey) public view returns (uint){                      
+        return (((frameKey - (uint(block.timestamp))) / period) + 1);
     }
 
     /// @notice Get lot struct
     /// @param frameKey Frame's timestamp
     /// @param lotKey lot's key 
     /// @return lot struct
-    function getLot(uint64 frameKey, uint lotKey) public view returns (Lot memory){                 
+    function getLot(uint frameKey, uint lotKey) public view returns (Lot memory){                 
         return frames[frameKey].lots[lotKey];
     }
 
     /// @notice Get no. of created lots in a frame 
     /// @param frameKey Frame's timestamp
     /// @return lot count
-    function getLotsCount(uint64 frameKey) public view returns (uint){                 
+    function getLotsCount(uint frameKey) public view returns (uint){                 
         return frames[frameKey].lotKeys.length;
     }
 
@@ -165,7 +165,7 @@ contract Market {
     /// @param frameKey Frame's timestamp
     /// @param index frameKeys array index 
     /// @return lotKey
-    function getLotKey(uint64 frameKey, uint index) public view returns (uint){                 
+    function getLotKey(uint frameKey, uint index) public view returns (uint){                 
         return frames[frameKey].lotKeys[index];
     }
 
@@ -173,7 +173,7 @@ contract Market {
     /// @notice Remaining tax of a lot owner get's returned after a sale (change of ownership). So the award amount isn't for sure untill the frame get's closed.
     /// @param frameKey Frame's timestamp
     /// @return award amount
-    function getRewardAmount(uint64 frameKey) public view returns (uint) {                         
+    function getRewardAmount(uint frameKey) public view returns (uint) {                         
         return frames[frameKey].rewardFund; 
     }
 
@@ -186,12 +186,12 @@ contract Market {
     /// @notice Create a frame
     /// @param timestamp In second.
     /// @return frameTimestamp Frame's key (timestamp that indicated the beggining of a frame).
-    function getOrCreateFrame(uint64 timestamp) public returns (uint64){
+    function getOrCreateFrame(uint timestamp) public returns (uint){
         //Get frame's timestamp - frame key
-        uint64 frameKey = clcFrameTimestamp(timestamp);
+        uint frameKey = clcFrameTimestamp(timestamp);
         //Check if frame exists
         if (frames[frameKey].state != SFrame.NULL) return frameKey;
-        require(frameKey >= clcFrameTimestamp(uint64(block.timestamp)), "CAN'T CREATE FRAME IN PAST");
+        require(frameKey >= clcFrameTimestamp(uint(block.timestamp)), "CAN'T CREATE FRAME IN PAST");
         //Add frame
         frames[frameKey].frameKey = frameKey;
         frames[frameKey].state = SFrame.OPENED;
@@ -204,7 +204,7 @@ contract Market {
     /// @param frameKey Frame's timestamp
     /// @param pairPrice Price is the trading pair's price
     /// @return lotKey
-    function getOrCreateLot(uint64 frameKey, uint pairPrice) internal returns(uint){
+    function getOrCreateLot(uint frameKey, uint pairPrice) internal returns(uint){
         uint lotKey = clcLotInterval(pairPrice);
         if (frames[frameKey].lots[lotKey].state != SLot.NULL) return lotKey;
         frames[frameKey].lots[lotKey].lotKey = lotKey;
@@ -212,11 +212,11 @@ contract Market {
         return lotKey;
     }
 
-    function clcTax(uint64 frameKey, uint acquisitionPrice) public view returns (uint tax) {             
-        uint dFrame = (clcFrameTimestamp(uint64(block.timestamp) + (period))) - (block.timestamp); 
+    function clcTax(uint frameKey, uint acquisitionPrice) public view returns (uint tax) {             
+        uint dFrame = (clcFrameTimestamp(uint(block.timestamp) + (period))) - (block.timestamp); 
         uint dFrameP = scalar * dFrame / (period);
         //Calculate tax from the proposed acquisition price 
-        uint64 numOfFramesLeft = clcFramesLeft(frameKey);
+        uint numOfFramesLeft = clcFramesLeft(frameKey);
         tax = acquisitionPrice * taxMarket / 100000 * numOfFramesLeft;
         uint dtax = acquisitionPrice * taxMarket / 100000;
         dtax = dtax * dFrameP / scalar;
@@ -229,8 +229,8 @@ contract Market {
     /// @param pairPrice Value is trading pair's price (to get correct lot key) 
     /// @param acqPrice New sell price is required to calculate tax
     /// @return cmpltAmnt complete price of a lot
-    function clcAmountToApprove(uint64 timestamp, uint pairPrice, uint acqPrice) public view returns (uint cmpltAmnt) {    
-        uint64 frameKey = clcFrameTimestamp(timestamp);         
+    function clcAmountToApprove(uint timestamp, uint pairPrice, uint acqPrice) public view returns (uint cmpltAmnt) {    
+        uint frameKey = clcFrameTimestamp(timestamp);         
         uint lotKey = clcLotInterval(pairPrice);
         uint tax = clcTax(frameKey, acqPrice);
         cmpltAmnt = frames[frameKey].lots[lotKey].acquisitionPrice + tax;
@@ -241,11 +241,11 @@ contract Market {
     /// @param timestamp Frame's timestamp get's calculated from this value
     /// @param pairPrice Is trading pair's price (to get correct lot key) 
     /// @param acqPrice New sell price is required to calculate tax
-    function buyLot(uint64 timestamp, uint pairPrice, uint acqPrice) public minTaxCheck(acqPrice) {
+    function buyLot(uint timestamp, uint pairPrice, uint acqPrice) public minTaxCheck(acqPrice) {
         //Get frameKey and lotKey values
-        uint64 frameKey =  getOrCreateFrame(timestamp);                          
+        uint frameKey =  getOrCreateFrame(timestamp);                          
         uint lotKey = getOrCreateLot(frameKey, pairPrice);
-        require(frameKey >= clcFrameTimestamp(uint64(block.timestamp)), "CAN'T BUY A LOT FROM THE PAST");
+        require(frameKey >= clcFrameTimestamp(uint(block.timestamp)), "CAN'T BUY A LOT FROM THE PAST");
         require(msg.sender != frames[frameKey].lots[lotKey].lotOwner, "ADDRESS ALREADY OWNS THE PARCEL");
 
         uint tax = clcTax(frameKey, acqPrice);
@@ -276,12 +276,13 @@ contract Market {
         emit LotUpdate(frameKey, lotKey);
     }
 
+    //TODO: users shouldn't be able to change prices of lots in the past
     /// @notice Owner can update lot's price. Has to pay additional tax, or leftover tax gets' returned to him if the new price is lower
     /// @param timestamp Frame's timestamp get's calculated from this value
     /// @param pairPrice Is trading pair's price (to get correct lot key) 
     /// @param acqPrice New acquisition price
-    function updateLotPrice(uint64 timestamp, uint pairPrice, uint acqPrice) public minTaxCheck(acqPrice) {
-        uint64 frameKey = getOrCreateFrame(timestamp); 
+    function updateLotPrice(uint timestamp, uint pairPrice, uint acqPrice) public minTaxCheck(acqPrice) {
+        uint frameKey = getOrCreateFrame(timestamp); 
         uint lotKey = getOrCreateLot(frameKey, pairPrice); 
         require(frames[frameKey].lots[lotKey].lotOwner == msg.sender, "MUST BE LOT OWNER");
 
@@ -307,8 +308,8 @@ contract Market {
     /// @notice Update trading pair's prices in the frame
     function updateFramePrices() public {
         uint tmp;
-        uint64 frameKey = clcFrameTimestamp(uint64(block.timestamp));
-        frames[frameKey].lastBlockNum = uint64(block.number);
+        uint frameKey = clcFrameTimestamp(block.timestamp);
+        frames[frameKey].lastBlockNum = block.number;
         //Correct price if outside settle interval
         if (block.timestamp >= ((frameKey + (period)) - (tReporting)) && frames[frameKey].oraclePrice0CumulativeStart == 0) {
             (frames[frameKey].oraclePrice0CumulativeStart, tmp, frames[frameKey].oracleTimestampStart) = UniswapV2OracleLibrary.currentCumulativePrices(address(uniswapPair));
@@ -325,24 +326,24 @@ contract Market {
     // /// @dev For developement purposes the new price is added as an input
     // /// @dev Final version should use uniswap's oracles
     // function updateFramePrices(uint PriceCumulative) public {
-    //     uint64 frameKey = clcFrameTimestamp(uint64(block.timestamp));
-    //     frames[frameKey].lastBlockNum = uint64(block.number);
+    //     uint frameKey = clcFrameTimestamp(uint(block.timestamp));
+    //     frames[frameKey].lastBlockNum = uint(block.number);
     //     //Correct price if outside settle interval
     //     if (block.timestamp >= ((frameKey + period) - tReporting) && frames[frameKey].oraclePrice0CumulativeStart == 0) {
     //         frames[frameKey].oraclePrice0CumulativeStart = PriceCumulative;
-    //         frames[frameKey].oracleTimestampStart = uint64(block.timestamp);
+    //         frames[frameKey].oracleTimestampStart = uint(block.timestamp);
     //         emit FrameUpdate(frameKey);
     //     }
     //     else if (block.timestamp >= ((frameKey + period) - tReporting)) {
     //         frames[frameKey].oraclePrice0CumulativeEnd = PriceCumulative;
-    //         frames[frameKey].oracleTimestampEnd = uint64(block.timestamp);
+    //         frames[frameKey].oracleTimestampEnd = uint(block.timestamp);
     //         emit FrameUpdate(frameKey);
     //     }
     // }
 
     /// @notice Close frame
     /// @param frameKey Frame's timestamp
-    function closeFrame(uint64 frameKey) public hasValidPrices(frameKey) {
+    function closeFrame(uint frameKey) public hasValidPrices(frameKey) {
         require(frames[frameKey].state == SFrame.OPENED, "FRAME NOT OPENED");
         require( (frameKey + period) <= block.timestamp, "FRAME END TIME NOT REACHED");
         frames[frameKey].state = SFrame.CLOSED;
@@ -365,7 +366,7 @@ contract Market {
 
     /// @notice Settle lot for a owner. Owner's share of the reward amount get's transfered to owner's account
     /// @param frameKey Frame's timestamp
-    function settleLot(uint64 frameKey) public {  
+    function settleLot(uint frameKey) public {  
         //Check if frame is closed
         require(frames[frameKey].state == SFrame.CLOSED, "FRAME NOT CLOSED");    
         //Calcualte the winning lot key
